Revert "different bot implementation"
This reverts commit 37804dac055ae0d3072bb40ca89ebd3812395ab4.
