+++ lib/console/chat/impl/slack.ex
@@ -1,10 +1,19 @@
 defmodule Console.Chat.Impl.Slack do
+  use Slack.Bot
   use Console.Chat.Impl
-  alias Console.Chat.Channel
+  alias Console.Chat.{Channel, Utils, Reference}
 
   @limit 1000
 
-  def child_spec(%ChatConnection{} = conn), do: {SlackBot, :start_link, [slack_args(conn)]}
+  def handle_event(_, %{"ts" => ts, "text" => text, "channel" => channel}, %Slack.Bot{user_id: id} = bot) do
+    case String.contains?(text, "<@#{id}>") do
+      true -> spawn_job(ts, text, channel, bot)
+      false -> :ok
+    end
+  end
+  def handle_event(_, _, _), do: :ok
+
+  def child_spec(%ChatConnection{} = conn), do: {Slack.Supervisor, :start_link, [slack_args(conn)]}
 
   def search_channels(%ChatConnection{type: :slack, configuration: %{slack: %{bot_token: token}}}, query)
       when is_binary(token) do
@@ -32,12 +41,31 @@ defmodule Console.Chat.Impl.Slack do
     }
   }} = conn) when is_binary(app_token) and is_binary(bot_token) do
     [
-      module: Console.Chat.Impl.Slack.Bot,
-      name: :"slack_bot_#{conn.id}",
-      config: [assigns: %{conn: conn}],
-      app_token: app_token,
       bot_token: bot_token,
+      bot: __MODULE__,
+      bot_assigns: %{conn: conn},
+      app_token: app_token,
+      channels: [types: ~w(public_channel private_channel)],
+      supervisor_args: [name: {:via, Registry, {Console.AI.Agents, {:slack_bot, conn.id}}}]
     ]
+  end
+
+  defp spawn_job(ts, text, channel, %Slack.Bot{assigns: %{conn: %ChatConnection{} = conn}, token: token}) do
+    with {:ok, name} <- fetch_channel(channel, token),
+         {:ok, _} <- Utils.handle_mention(%Reference{id: ts, text: text}, %Reference{id: channel, text: name}, conn) do
+      :ok
+    else
+      :ok -> :ok
+      err -> Logger.error("failed to spawn job: #{inspect(err)}")
+    end
+  end
+  defp spawn_job(_, _, _, _), do: :ok
+
+  defp fetch_channel(id, token) do
+    case Slack.API.get("conversations.info", token, channel: id) |> IO.inspect(label: "slack conversation") do
+      {:ok, %{"ok" => true, "channel" => %{"name" => name}}} -> {:ok, name}
+      result -> {:error, "Failed to fetch channel: #{inspect(result)}"}
+    end
   end
 
   defp filter(channels, query) when is_binary(query) do
