+++ mix.exs
@@ -87,7 +87,6 @@ defmodule Console.MixProject do
       {:crontab, "~> 1.1"},
       {:tentacat, "~> 2.0", git: "https://github.com/pluralsh/tentacat.git", ref: "c2c4c30b113c70b2ff2cd674fc04492b87d04100"},
       {:slack_elixir, git: "https://github.com/pluralsh/slack_elixir.git", ref: "a55bf71744bd7b80ecebe75a60e54953cf8a8514"},
-      {:slack_bot_ws, "~> 0.1.0-rc.2"},
       {:absinthe_client, "~> 0.1.0"},
       {:postgrex, ">= 0.0.0"},
       {:grpc, "~> 0.11"},
